+++ lib/src/core/digit.dart
@@ -1,26 +1,89 @@
+import 'package:formdator/formdator.dart';
+
 /// Digit - any one of the ten numbers 0 to 9.
 ///
 /// For example, the values "1" and "123" are valid; whereas the values "A",
 /// "12A", "123.34", "9.99" are invalid.
 ///
-/// Blank field - null value - is a valid input.
+/// Blank field — null or empty — is a valid input.
 ///
 /// If the field is mandatory, see [Req].
 class Digit {
-  /// Validates an optional digit-only field.
+  /// Constrains input data to digits [0-9].
   ///
-  /// [non] is the error message in case a non-digit character is found in the
-  /// input data; defaults to 'non-digit character'.
-  const Digit({String? non}) : _nonDigit = non ?? 'non-digit character';
+  /// [non] the error message if a non-digit character is found; the default
+  /// value is 'non-digit character'.
+  Digit({String? non}) : _valDig = _DigitImpl(non, const Ok());
 
-  final String _nonDigit;
+  /// Constrains input data to digits [0-9] and limits the length (number of
+  /// digits) to [len] digits.
+  ///
+  /// [len] the number of digits; it must be > 0.
+  /// [non] the error message if a non-digit character is found.
+  /// [diff] the error message if the number of digits is different from [len].
+  Digit.len(int len, {String? non, String? diff})
+      : assert(len > 0),
+        _valDig = _DigitImpl(non, Len(len, error: diff));
 
-  static late final _digitsOnly = RegExp(r'^[0-9]+$');
+  /// Constrains thinput data to digits [0-9] and limits the length (number of
+  /// digits) to a minimum of [min] digits.
+  ///
+  /// [min] the minimum number of digits; it must be > 0.
+  /// [non] the error message if a non-digit character is found.
+  /// [short] the error message if the number of digits is less than [max].
+  Digit.min(int min, {String? non, String? short})
+      : assert(min > 0),
+        _valDig = _DigitImpl(non, Len.min(min, short: short));
+
+  /// Constrains input data to digits [0-9] and limits the length (number of
+  /// digits) to a maximum of [max] digits.
+  ///
+  /// [max] the maximum number of digits; it must be > 0.
+  /// [non] the error message if a non-digit character is found.
+  /// [long] the error message if the number of digits is greater than [max].
+  Digit.max(int max, {String? non, String? long})
+      : assert(max > 0),
+        _valDig = _DigitImpl(non, Len.max(max, long: long));
+
+  /// Constrains input data to digits [0-9] and limits the length (number of
+  /// digits) within the range [min–max].
+  ///
+  /// [min] the minimum number of digits; it must be > 0 and < [max].
+  /// [max] the maximum number of digits; it must be > 0 and > [min].
+  /// [non] the error message if a non-digit character is found.
+  /// [short] the error message if the number of digits is less than [min].
+  /// [long] the error message if the number of digits is greater than [max].
+  Digit.range(int min, int max, {String? non, String? short, String? long})
+      : assert(min > 0),
+        assert(max > 0),
+        assert(max > min),
+        _valDig = _DigitImpl(
+          non,
+          Len.range(min, max, short: short, long: long),
+        );
+
+  // the digit-only validation logic.
+  final ValStr _valDig;
 
   /// Valid - returns null - if [digits] has only digits or is null.
-  String? call(String? digits) =>
-      (digits == null || _isNumeric(digits)) ? null : _nonDigit;
+  String? call(String? digits) => _valDig(digits);
+}
 
-  /// Checks whether [value] contains only digits.
-  bool _isNumeric(String value) => _digitsOnly.hasMatch(value);
+class _DigitImpl {
+  /// Ctor. [non] non-digit error message; [extra] additional validation step.
+  _DigitImpl(String? non, ValObj extra)
+      : _nonDigit = non ?? 'non-digit character',
+        _extra = extra;
+
+  final String _nonDigit;
+  // extra validation step.
+  final ValObj _extra;
+
+  static final _digitsOnly = RegExp(r'^[0-9]+$');
+
+  String? call(String? input) =>
+      // (input == null || _digitsOnly.hasMatch(input)) ? null : _non;
+      (input != null && !_digitsOnly.hasMatch(input))
+          ? _nonDigit
+          : _extra(input);
 }